fix order checkout

+++ src/Abstractions/IOrderItemRepositories.cs
@@ -6,7 +6,7 @@ namespace sda_onsite_2_csharp_backend_teamwork_The_countryside_developers
 
         public IEnumerable<OrderItem> FindAll();
 
-        public OrderItem? UpdateOne(Guid orderItemId, int newQuantity, decimal newTotalPrice);
+        public OrderItem? UpdateOne(Guid orderItemId, int newQuantity, double newTotalPrice);
