+++ src/OrderItem/Entities/OrderItem.cs
@@ -10,7 +10,7 @@ namespace sda_onsite_2_csharp_backend_teamwork_The_countryside_developers
         public Guid ProductId { get; set; }
         public int Quantity { get; set; }
 
-        public decimal TotalPirce { get; set; }
+        public double TotalPirce { get; set; }
