+++ src/OrderItem/Repositories/OrderItemRepository.cs
@@ -23,7 +23,7 @@ namespace sda_onsite_2_csharp_backend_teamwork_The_countryside_developers
         }
 
 
-        public OrderItem? UpdateOne(Guid orderItemId, int newQuantity, decimal newTotalPrice)
+        public OrderItem? UpdateOne(Guid orderItemId, int newQuantity, double newTotalPrice)
         {
 
             OrderItem? itemToUpdate = _orderitems.FirstOrDefault(item => item.Id == orderItemId);
